missing cast
git-svn-id: https://svn.r-project.org/R/trunk@60790 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/internet.c
@@ -344,7 +344,7 @@ SEXP Rsockwrite(SEXP ssock, SEXP sstring)
         error("invalid 'socket' argument");
     int sock = asInteger(ssock), start = 0, end, len;
     char *buf = (char *)translateChar(STRING_ELT(sstring, 0)), *abuf[1];
-    end = len = strlen(buf);
+    end = len = (int)strlen(buf);
     abuf[0] = buf;
     if (!initialized)
         internet_Init();
